requirements/test.txt
coveralls
pytest
pytest-cov
pytest-datadir
pytest-socket
coverage!=6.0.*,!=6.1,!=6.1.1,<7.0,>=4.1
python-dotenv>=0.19.0
rich>=12.0.0